refactor: convert auth policy components into typescript
