+++ studio/components/interfaces/Authentication/Policies/PolicyEditor/PolicyName.tsx
@@ -1,0 +1,30 @@
+import { FC } from 'react'
+import { Input } from '@supabase/ui'
+
+interface Props {
+  name: string
+  onUpdatePolicyName: (name: string) => void
+}
+
+const PolicyName: FC<Props> = ({ name = '', onUpdatePolicyName }) => {
+  return (
+    <div className="flex space-x-12">
+      <div className="flex w-1/3 flex-col space-y-2">
+        <label className="text-scale-1100 text-base" htmlFor="policy-name">
+          Policy name
+        </label>
+        <p className="text-scale-900 text-sm">A descriptive name for your policy</p>
+      </div>
+      <div className="relative w-2/3">
+        <Input
+          id="policy-name"
+          value={name}
+          onChange={(e) => onUpdatePolicyName(e.target.value)}
+          actions={<span className="text-scale-900 mr-3 text-sm">{name.length}/63</span>}
+        />
+      </div>
+    </div>
+  )
+}
+
+export default PolicyName
