+++ studio/components/interfaces/Authentication/Policies/PolicySelection.tsx
@@ -1,62 +1,63 @@
+import { FC } from 'react'
 import { IconEdit, IconGrid, Modal } from '@supabase/ui'
 import CardButton from 'components/ui/CardButton'
 
-const PolicySelection = ({
+interface Props {
+  description: string
+  onViewTemplates: () => void
+  onViewEditor: () => void
+}
+
+const PolicySelection: FC<Props> = ({
   description = '',
   onViewTemplates = () => {},
   onViewEditor = () => {},
 }) => {
   return (
     <Modal.Content>
-      <div className="py-8 space-y-6">
+      <div className="space-y-6 py-8">
         <div>
-          <p className="text-sm text-scale-1100">{description}</p>
+          <p className="text-scale-1100 text-sm">{description}</p>
         </div>
-        <div className="grid grid-cols-1 lg:grid-cols-1 gap-2">
+        <div className="grid grid-cols-1 gap-2 lg:grid-cols-1">
           <CardButton
-            hoverable
-            className="cursor-pointer py-4"
-            onClick={onViewTemplates}
-            // imgUrl={'/img/policy-template.svg'}
             title="Get started quickly"
             description="Create a policy from a template"
             icon={
               <div className="flex">
                 <div
                   className="
-                  w-8 h-8 bg-scale-1200 text-scale-100
-                  rounded
-                  flex items-center justify-center  
+                  bg-scale-1200 text-scale-100 flex h-8
+                  w-8
+                  items-center justify-center rounded  
                 "
                 >
                   <IconGrid size={14} strokeWidth={2} />
                 </div>
               </div>
             }
+            onClick={onViewTemplates}
           />
-
           <CardButton
-            hoverable
-            className="cursor-pointer py-4"
-            onClick={onViewEditor}
             title="For full customization"
             description="Create a policy from scratch"
             icon={
               <div className="flex">
                 <div
                   className="
-                  w-8 h-8 bg-scale-1200 text-scale-100
-                  rounded
-                  flex items-center justify-center  
+                  bg-scale-1200 text-scale-100 flex h-8
+                  w-8
+                  items-center justify-center rounded  
                 "
                 >
                   <IconEdit size={14} strokeWidth={2} />
                 </div>
               </div>
             }
+            onClick={onViewEditor}
           />
         </div>
-        <p className="text-sm text-scale-1100">
+        <p className="text-scale-1100 text-sm">
           Not sure what policies are? Check out our resources{' '}
           <a
             target="_blank"
