+++ studio/components/interfaces/Authentication/Policies/PolicyTemplates/index.tsx
@@ -1,0 +1,47 @@
+import { Button } from '@supabase/ui'
+import { FC, useState } from 'react'
+import { isEmpty } from 'lodash'
+
+import TemplatesList from './TemplatesList'
+import TemplatePreview from './TemplatePreview'
+
+interface Props {
+  templates: any[]
+  templatesNote: string
+  onUseTemplate: (template: any) => void
+}
+
+const PolicyTemplates: FC<Props> = ({
+  templates = [],
+  templatesNote = '',
+  onUseTemplate = () => {},
+}) => {
+  const [selectedTemplate, setSelectedTemplate] = useState(templates[0])
+  return (
+    <div>
+      <div className="dark:border-dark flex justify-between border-t">
+        <TemplatesList
+          templatesNote={templatesNote}
+          templates={templates}
+          selectedTemplate={selectedTemplate}
+          setSelectedTemplate={setSelectedTemplate}
+        />
+        <TemplatePreview selectedTemplate={selectedTemplate} />
+      </div>
+      <div className="dark:border-dark flex w-full items-center justify-end gap-3 border-t px-6 py-4">
+        <span className="text-scale-900 text-sm">
+          This will override any existing code you've written
+        </span>
+        <Button
+          type="primary"
+          disabled={isEmpty(selectedTemplate)}
+          onClick={() => onUseTemplate(selectedTemplate)}
+        >
+          Use this template
+        </Button>
+      </div>
+    </div>
+  )
+}
+
+export default PolicyTemplates
