+++ studio/components/interfaces/Authentication/Policies/index.tsx
@@ -1,0 +1,6 @@
+import PolicyEditorModal from './PolicyEditorModal'
+import PolicyTableRow from './PolicyTableRow'
+import PolicySelection from './PolicySelection'
+import PolicyTemplates from './PolicyTemplates'
+
+export { PolicyEditorModal, PolicyTableRow, PolicySelection, PolicyTemplates }
