+++ studio/components/to-be-cleaned/Storage/StoragePolicies/StoragePolicies.js
@@ -11,7 +11,7 @@ import StoragePoliciesEditPolicyModal from './StoragePoliciesEditPolicyModal'
 import ConfirmModal from 'components/ui/Dialogs/ConfirmDialog'
 import { useStorageStore } from 'localStores/storageExplorer/StorageExplorerStore'
 
-import PolicyEditorModal from 'components/to-be-cleaned/Auth/PolicyEditorModal'
+import { PolicyEditorModal } from 'components/interfaces/Authentication/Policies'
 
 const StoragePolicies = () => {
   const { ui, meta } = useStore()
@@ -140,7 +140,7 @@ const StoragePolicies = () => {
   }
 
   return (
-    <div className="flex flex-col w-full min-h-full">
+    <div className="flex min-h-full w-full flex-col">
       <h4 className="text-xl">Storage policies</h4>
       <p className="text-scale-1100">
         Safeguard your files with policies that define the operations allowed for your users at the
@@ -148,11 +148,11 @@ const StoragePolicies = () => {
       </p>
 
       {!loaded ? (
-        <div className="h-full flex items-center justify-center">
+        <div className="flex h-full items-center justify-center">
           <IconLoader className="animate-spin" size={16} />
         </div>
       ) : (
-        <div className="space-y-4 mt-4">
+        <div className="mt-4 space-y-4">
           {buckets.length === 0 && <StoragePoliciesPlaceholder />}
 
           {/* Sections for policies grouped by buckets */}
@@ -176,7 +176,7 @@ const StoragePolicies = () => {
             )
           })}
 
-          <div className="w-full border-b border-gray-600 !mb-4" />
+          <div className="!mb-4 w-full border-b border-gray-600" />
           <Typography.Text className="opacity-50">
             You may also write policies for the tables under the storage schema directly for greater
             control
